Fixed some types for compliance.
git-svn-id: svn://svn.code.sf.net/p/chibios/svn/trunk@15281 27425a3e-05d8-49a3-a47f-9c15f0e5edd8

+++ os/vfs/include/vfsnodes.h
@@ -62,7 +62,7 @@ typedef struct vfs_driver vfs_driver_c;
 /**
  * @brief   Type of a file offset.
  */
-typedef uint32_t vfs_offset_t;
+typedef off_t vfs_offset_t;
 
 /**
  * @brief   Type of a node attributes.
